Offset key: fixed storing offsets and lengths in 2 bytes instead of 4 bytes.

+++ src/Starcounter/Query/Execution/Enumerators/Aggregation.cs
@@ -373,7 +373,7 @@ internal class Aggregation : ExecutionEnumerator, IExecutionEnumerator
     /// <summary>
     /// Saves the underlying enumerator state.
     /// </summary>
-    public unsafe Int32 SaveEnumerator(Byte* keysData, Int32 globalOffset, Boolean saveDynamicDataOnly)
+    public unsafe UInt16 SaveEnumerator(Byte* keysData, UInt16 globalOffset, Boolean saveDynamicDataOnly)
     {
         return enumerator.SaveEnumerator(keysData, globalOffset, saveDynamicDataOnly);
     }
